requirements.txt
django
djangorestframework
django-filter
djangorestframework-simplejwt
drf-yasg
gunicorn
psycopg2-binary
dj-database-url
